Use the endpoint.Access type
