+++ config/config_test.go
@@ -145,82 +145,6 @@ func TestBatchAbsentIsTrue(t *testing.T) {
 	assert.True(t, v)
 }
 
-func TestAccessConfig(t *testing.T) {
-	type accessTest struct {
-		Access        string
-		PrivateAccess bool
-	}
-
-	tests := map[string]accessTest{
-		"":            {"none", false},
-		"basic":       {"basic", true},
-		"BASIC":       {"basic", true},
-		"private":     {"basic", true},
-		"PRIVATE":     {"basic", true},
-		"invalidauth": {"invalidauth", true},
-	}
-
-	for value, expected := range tests {
-		cfg := NewFrom(Values{
-			Git: map[string]string{
-				"lfs.url":                        "http://example.com",
-				"lfs.http://example.com.access":  value,
-				"lfs.https://example.com.access": "bad",
-			},
-		})
-
-		if access := cfg.Access("download"); access != expected.Access {
-			t.Errorf("Expected Access() with value %q to be %v, got %v", value, expected.Access, access)
-		}
-		if access := cfg.Access("upload"); access != expected.Access {
-			t.Errorf("Expected Access() with value %q to be %v, got %v", value, expected.Access, access)
-		}
-
-		if priv := cfg.PrivateAccess("download"); priv != expected.PrivateAccess {
-			t.Errorf("Expected PrivateAccess() with value %q to be %v, got %v", value, expected.PrivateAccess, priv)
-		}
-		if priv := cfg.PrivateAccess("upload"); priv != expected.PrivateAccess {
-			t.Errorf("Expected PrivateAccess() with value %q to be %v, got %v", value, expected.PrivateAccess, priv)
-		}
-	}
-
-	// Test again but with separate push url
-	for value, expected := range tests {
-		cfg := NewFrom(Values{
-			Git: map[string]string{
-				"lfs.url":                           "http://example.com",
-				"lfs.pushurl":                       "http://examplepush.com",
-				"lfs.http://example.com.access":     value,
-				"lfs.http://examplepush.com.access": value,
-				"lfs.https://example.com.access":    "bad",
-			},
-		})
-
-		if access := cfg.Access("download"); access != expected.Access {
-			t.Errorf("Expected Access() with value %q to be %v, got %v", value, expected.Access, access)
-		}
-		if access := cfg.Access("upload"); access != expected.Access {
-			t.Errorf("Expected Access() with value %q to be %v, got %v", value, expected.Access, access)
-		}
-
-		if priv := cfg.PrivateAccess("download"); priv != expected.PrivateAccess {
-			t.Errorf("Expected PrivateAccess() with value %q to be %v, got %v", value, expected.PrivateAccess, priv)
-		}
-		if priv := cfg.PrivateAccess("upload"); priv != expected.PrivateAccess {
-			t.Errorf("Expected PrivateAccess() with value %q to be %v, got %v", value, expected.PrivateAccess, priv)
-		}
-	}
-
-}
-
-func TestAccessAbsentConfig(t *testing.T) {
-	cfg := NewFrom(Values{})
-	assert.Equal(t, "none", cfg.Access("download"))
-	assert.Equal(t, "none", cfg.Access("upload"))
-	assert.False(t, cfg.PrivateAccess("download"))
-	assert.False(t, cfg.PrivateAccess("upload"))
-}
-
 func TestLoadValidExtension(t *testing.T) {
 	cfg := NewFrom(Values{
 		Git: map[string]string{},
